% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{default_config_path}
\alias{default_config_path}
\title{Default Path to Search for Local Configuration}
\usage{
default_config_path()
}
\description{
This function returns a default for the local config file.
This is given by the following:\if{html}{\out{<div class="r">}}\preformatted{file.path(here::here(), "local.yaml")
}\if{html}{\out{</div>}}

That is:

\emph{Assuming} \strong{.} \emph{refers to the root of the current Rproject:}

"./local.yaml"
}
