% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{describe_list_function}
\alias{describe_list_function}
\title{Describe a function that produces a list}
\usage{
describe_list_function(fn)
}
\description{
Describe a function that produces a list
}
\keyword{internal}
