% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_access.R
\name{export_deidentified_notes}
\alias{export_deidentified_notes}
\title{Export deidentified notes}
\usage{
export_deidentified_notes(dir_out, cache)
}
\arguments{
\item{dir_out}{the path to the directory under which to save the targets}

\item{cache}{the \link[drake:drake_cache]{drake::drake_cache} object from which the values will be
retrieved}
}
\description{
Export deidentified notes as rds from the cache to a directory
This function is to be used to provide deidentified inputs to
the pipeline (Which were previously outputs of the pipeline),
in order to fully remove identifiable data from the pipeline.
}
