% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_access.R
\name{export_target_set}
\alias{export_target_set}
\title{Export all partitions of a table to file}
\usage{
export_target_set(target_set_name, dir_out, cache)
}
\arguments{
\item{target_set_name}{a \strong{single} quoted \code{"string"}, providing the name of the
target, or cluster of targets.}

\item{dir_out}{the path to the directory under which to save the targets}

\item{cache}{the \link[drake:drake_cache]{drake::drake_cache} object from which the values will be
retrieved}
}
\description{
Save all partitions of a particular table to a directory.
Presently saves using the \code{qs} format, as this is meant to be fast (ooh).
For now I'm going to default it to using 4 threads
}
