% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_cache}
\alias{get_cache}
\title{Load the Drake Cache for the Project}
\usage{
get_cache(root = NULL)
}
\arguments{
\item{root}{the project root (Top level directory containing "pipeline" directory), defaults to
whatever root is currently set by \code{\link[=get_portrpath]{get_portrpath()}}}
}
\description{
Given the root directory of the project, return the cache to be used by
drake.
}
\details{
The cache will be of the SQLite implementation, to improve performance on
the lustre file system of the HPC.

This isn't intinsically threadsafe, so in \code{drake::config()}, be sure to
set \code{caching = "master"}, as per
\href{https://ropenscilabs.github.io/drake-manual/storage.html#database-caches}{drake's documentation}
}
\seealso{
Other paths: 
\code{\link{get_paths}()}
}
\concept{paths}
