% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_paths}
\alias{get_paths}
\title{Get a list of the main paths to be used}
\usage{
get_paths(config_path = NULL)
}
\description{
Get a list of the main paths to be used
}
\seealso{
Other paths: 
\code{\link{get_cache}()}
}
\concept{paths}
