% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_project_metadata}
\alias{get_project_metadata}
\title{Get Metadata for a Particular Project}
\usage{
get_project_metadata(root = NULL)
}
\arguments{
\item{root}{the project root (Top level directory containing "pipeline" directory), defaults to
whatever root is currently set by \code{\link[=get_portrpath]{get_portrpath()}}}
}
\description{
For a project whose data is rooted at \code{root}, get the associated metadata.
If none exists, query for it interactively.

An example metadata file is as follows:
}
\section{Metadata Example}{
NA
}

\seealso{
Other metadata: 
\code{\link{metadata_template}()}
}
\concept{metadata}
