% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_access.R
\name{map_fetch}
\alias{map_fetch}
\title{Map over each target in the set before combining with the rest}
\usage{
map_fetch(.target_set_name, .f, ..., .cache = NULL)
}
\arguments{
\item{.target_set_name}{the prefix for each target eg. \code{"journey_analysis_base"}
matches \code{"journey_analysis_base_1"}, \code{"journey_analysis_base_2"}, \verb{_3}, \verb{_4} and so on.}

\item{.f}{a function to apply to each sharded input after it is fetched, before it combined with
the rest of the inputs}

\item{.cache}{the \code{storr} cache from which the targets are fetched.}
}
\description{
Map over each target in the set before combining with the rest
}
