% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_template}
\alias{metadata_template}
\title{An example Metadata file for a dataset}
\usage{
metadata_template()
}
\description{
The template is as follows
}
\section{Metadata Example}{
\if{html}{\out{<div class="yaml">}}\preformatted{title: title_of_particular_workflow
desc: "A short description of the particular workflow.\\nThis is a good place to describe
  problems with the raw data, its \\nextract period, and so on."
log_path: relative/path/from/Rproject/root/title_of_particular_workflow

}\if{html}{\out{</div>}}
}

\seealso{
Other metadata: 
\code{\link{get_project_metadata}()}
}
\concept{metadata}
