% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty_string}
\alias{pretty_string}
\title{Log an object in its native print representation (with newlines)}
\usage{
pretty_string(x, ...)
}
\arguments{
\item{x}{the object to be displayed}
}
\description{
Log an object in its native print representation (with newlines)
}
\keyword{internal}
