% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_paths}
\alias{get_paths}
\title{Get a list of the main paths to be used}
\usage{
get_paths(config_path = NULL)
}
\arguments{
\item{config_path}{the path to the configuration \code{yaml} file. If not
specified, this will be given by \code{getOption("pipetree.config")},
and if this is unset, will be given by \code{\link[=default_config_path]{default_config_path()}}.
(See \strong{Default Config Path} section)}
}
\description{
Get a list of the main paths to be used
}
\section{Default Config Path}{

This is given by the following:\if{html}{\out{<div class="r">}}\preformatted{file.path(here::here(), "local.yaml")
}\if{html}{\out{</div>}}

That is:

\emph{Assuming} \strong{.} \emph{refers to the root of the current Rproject:}

"./local.yaml"
}

\seealso{
Other paths: 
\code{\link{get_cache}()}
}
\concept{paths}
