% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_portrpath}
\alias{get_portrpath}
\title{Get the path Configuration object}
\usage{
get_portrpath(config_path = NULL)
}
\description{
Get the \code{PortrPath} object for configuring paths, profiles etc.
}
\details{
See
\href{https://mstr3336.github.io/portrpaths/articles/configuring-your-paths.html}{portrpaths::PortrPath}
for more info.
}
