% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_project_metadata}
\alias{get_project_metadata}
\title{Get Metadata for a Particular Project}
\usage{
get_project_metadata(root = NULL, interactive = TRUE)
}
\arguments{
\item{root}{the project root (Top level directory containing "pipeline" directory), defaults to
whatever root is currently set by \code{\link[=get_portrpath]{get_portrpath()}}}

\item{interactive}{use this for noninteractive scripts to handle missing
metadata without user input}
}
\description{
For a project whose data is rooted at \code{root}, get the associated metadata.
If none exists, query for it interactively.

An example metadata file is as follows:
}
\section{Metadata Example}{
\if{html}{\out{<div class="yaml">}}\preformatted{title: title_of_particular_workflow
desc: "A short description of the particular workflow.\\nThis is a good place to describe
  problems with the raw data, its \\nextract period, and so on."
log_path: relative/path/from/Rproject/root/title_of_particular_workflow

}\if{html}{\out{</div>}}
}

\seealso{
Other metadata: 
\code{\link{metadata_template}()}
}
\concept{metadata}
