% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_template}
\alias{metadata_template}
\title{An example Metadata file for a dataset}
\usage{
metadata_template()
}
\description{
The template is as follows

\preformatted{}
}
\section{Metadata Example}{
NA
}

\seealso{
Other metadata: 
\code{\link{get_project_metadata}()}
}
\concept{metadata}
